e2e fixes
